\name{atmi-package}
\Rdversion{1.1}
\alias{atmi-package}
\alias{atmi}
\docType{package}
\title{
Analysis of technical market indicators
}
\description{
Analysis and usage of the trading rules, which are based on technical market indicators as well as on the time series analysis.
}
\details{
\tabular{ll}{
Package: \tab atmi\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-10-05\cr
Depends: \tab R (>= 2.9.1), TTR\cr
License: \tab GPL (>=2)\cr
}
The package includes four main functions: \code{\link{atmimomentum}}, \code{\link{atmirsi}} and \code{\link{atmimacd}}. As well as seven auxiliary functions: \code{\link{getsymbols}}, \code{\link{getdata}}, \code{\link{intersection}}, \code{\link{stoploss}}, \code{\link{superfluous_filter}}, \code{\link{performance}} and \code{\link{naivpredict}}.
}
\author{
Waldemar Kemler, Peter Schaffner, <atmi.info@gmail.com>

Maintainer: Waldemar Kemler, Peter Schaffner, <atmi.info@gmail.com>
}

\keyword{ ts }
